+++ wemake_python_styleguide/logic/tree/attributes.py
@@ -49,3 +49,12 @@ def is_foreign_attribute(node: AnyVariableDef) -> bool:
     # So, we change the strictness of this rule,
     # based on the attribute source.
     return node.value.id not in SPECIAL_ARGUMENT_NAMES_WHITELIST
+
+
+def is_special_attr(node: ast.AST) -> bool:
+    """Finds attributes that are assigned to `self`, `cls`, etc."""
+    if not isinstance(node, ast.Attribute):
+        return False
+    if not isinstance(node.value, ast.Name):
+        return False
+    return node.value.id in SPECIAL_ARGUMENT_NAMES_WHITELIST