+++ wemake_python_styleguide/visitors/ast/attributes.py
@@ -5,10 +5,10 @@ from typing_extensions import final
 from wemake_python_styleguide.compat.aliases import FunctionNodes
 from wemake_python_styleguide.constants import (
     ALL_MAGIC_METHODS,
-    SPECIAL_ARGUMENT_NAMES_WHITELIST,
 )
 from wemake_python_styleguide.logic import nodes
 from wemake_python_styleguide.logic.naming import access
+from wemake_python_styleguide.logic.tree import attributes
 from wemake_python_styleguide.violations.base import ASTViolation
 from wemake_python_styleguide.violations.best_practices import (
     ProtectedAttributeViolation,
@@ -37,10 +37,7 @@ class WrongAttributeVisitor(BaseNodeVisitor):
         node: ast.Attribute,
         exception: type[ASTViolation],
     ) -> None:
-        if (
-            isinstance(node.value, ast.Name)
-            and node.value.id in SPECIAL_ARGUMENT_NAMES_WHITELIST
-        ):
+        if attributes.is_special_attr(node):
             return
 
         if isinstance(node.value, ast.Call) and self._is_super_called(